==> notebooks/02_data_cleaning_feature_engineering.ipynb <==
from bs4 import BeautifulSoup
import requests
import pandas as pd
import json
import numpy as np
from itertools import zip_longest
import re


flip = pd.read_csv("data/raw/Flipcart.csv")


flip.head()

flip.columns

flip.drop("Unnamed: 0",axis=1,inplace=True)

flip.describe()

flip.info()

flip["Original_Price"] = flip["Original_Price"].str.replace("₹","")
flip["Discounted_Price"] = flip["Discounted_Price"].str.replace("₹","")
flip["Original_Price"] = flip["Original_Price"].str.replace(",","")
flip["Discounted_Price"] = flip["Discounted_Price"].str.replace(",","")


flip.head()

flip["Off_Upto"] = flip["Off_Upto"].str[:2]

flip["Name"] = flip["Name"].str.replace(r"\s*\(.*?\)", " ", regex=True).str.strip()


flip.head()

flip.isnull().sum()

flip["Original_Price"] = flip["Original_Price"].astype(float)
flip["Discounted_Price"] = flip["Discounted_Price"].astype(float)

flip.dtypes

# flip["Off_Upto"] = flip["Off_Upto"].astype(float)

flip["Off_Upto"] = flip["Off_Upto"].str.replace("%", "", regex=False)


flip["Off_Upto"] = flip["Off_Upto"].astype(float)

flip.dtypes

for i in range(len(flip["Name"])):
   print(flip["Name"].str.split(" ")[i])

comp = []
gen = []

def generation(name):
    for g in ["6G", "5G", "4G", "3G", "2G", "1G"]:
        if g in name:
            return g
    return np.nan
 
flip["Company"] = flip["Name"].str.split().str[0]
flip["Gen"] = flip["Name"].apply(generation)





flip.head()

flip["Off_Upto"]=flip["Off_Upto"].fillna(0)

flip.sample(5)

flip.isna().sum()

flip.drop("Gen",axis=1,inplace=True)

flip["Original_Price"].isna().sum()

flip["Original_Price"] = flip["Original_Price"].fillna(flip["Original_Price"].mean())

flip["Original_Price"].isna().sum()

flip["Display"].unique()    

def display_type(s):
    if(re.search(r'\)\s*(.*)$', s)):
        return re.search(r'\)\s*(.*)$', s).group(1)
    else:
        return np.nan
    
flip["Display_Type"] = flip["Display"].apply(display_type)

flip["Display_Type"].isna().sum()

def display_size(s):
    if(float(re.search(r'\(([\d.]+)', s).group(1))):
        return float(re.search(r'\(([\d.]+)', s).group(1))
    else:
        np.nan
flip["Display_Size"] = flip["Display"].apply(display_size)
    

flip["Display_Size"]

flip.drop("Display",axis=1,inplace=True)

flip

flip["Memory"].value_counts()

flip["Ram"] = flip["Memory"].str.split("|").str[0]


flip.loc[flip["Ram"].str.contains("ROM", na=False), "Ram"] = np.nan


mask = flip["Ram"].str.contains("MB", na=False)

flip.loc[mask, "Ram"] = (
    flip.loc[mask, "Ram"]
    .str.replace("MB RAM", "", regex=False)
    .str.strip()
    .astype(float)
    .div(1024)
    .astype(object)
)


flip["Ram"]=flip["Ram"].str.replace("GB RAM ","")



flip["Ram"] = flip["Ram"].str.replace(
    r'Expandable\s+Upto\s+\d+\s*GB',
   '',
    regex=True,
    case=False
)
flip["Ram"] = flip["Ram"].replace(r'^\s*$', np.nan, regex=True)

flip["Ram"].isna().sum()

flip["Ram"] = flip["Ram"].astype(float)

flip["Ram"]=flip["Ram"].fillna(0)

flip

flip["Rom"] = flip["Memory"].str.extract(r'\|\s*(.*)')



flip["Rom"] = (
    flip["Rom"]
    .str.replace(
        r'\|\s*Expandable\s+Upto\s+\d+\s*(GB|TB)',
        '',
        regex=True,
        case=False
    )
    .str.replace(r'\|\s*$', '', regex=True)  # remove trailing |
    .str.strip()
    .replace('', np.nan)
)

flip["Rom"]=flip["Rom"].str.replace("MB ROM","")



flip["Rom"] = flip["Rom"].fillna(0)


flip["Rom"] = flip["Rom"].astype(float)


flip.drop("Memory",axis=1,inplace=True)

flip.columns

flip.drop("Camera",axis=1,inplace=True)

flip.drop("Off_Upto",axis=1,inplace=True)


flip=flip[(flip["Rom"]!=0)&(flip["Ram"]!=0)]

flip["Ram"].value_counts()

flip["Rom"].value_counts()

flip.shape

flip.columns

flip = flip.reset_index(drop=True)


flip

flip.to_csv("data/clean/cleaned_flipkart.csv", index=False)
